ai-foundry-agents/requirements.txt
python-dotenv
openai
azure-ai-projects
azure-identity
agent-framework-azure-ai
azure-ai-documentintelligence

